Build up the emulator infrastructure
More of the basic infrastructure is now in place.
Command execution and command completion notifications
are present and working.

There is a new way to define objects that provides a way
for each object to define it's own custom data types for
configuration, command execution, and command completion.

The bulk of the diff, however, is formatting changes to
get the correct Emacs c-mode working.