Fix timing in object update thread
The object update thread recorded the time it took to
execute all of the object update callbacks and use that
to compute how long it needed to sleep in order to
complete one controller clock tick.

However, it did not take into account how long the entire
loop took and therefore was providing inaccurate timing to
the objects.

This change reworks how the loop timing is recorded, computed
and reported to the objects.