Add heater object
This commit adds a heater object that can be used to
emulate a heating element on a printer. It also adds
some definitions for a thermistor object. The hope was
to use the thermistor object for the heater object
since, at their core, they will be almost the same.

However, it does not seem to be possible to
programmatically get the field types of inherited
ctype.Structure structures.

This commit also includes a couple of tools to test
various ramp algorithms, as well as some datasheets for
various thermistor types.