Add object initialization code path
Previously, objects initialized themselves during
the object creation calls. That works most of the
time.

However, for objects that need to reference other
objects, the lookup will fail if the referenced
object has not been created yet.

This puts a requirement on the configuration file
that objects that reference other objects be defined
after the referenced objects. This behavior is not
user-friendly and creates potential for errors.

Instead of putting a requirement on the config file
structure, add a separate object initialization code
path. All objects defined in the config file are
first created. After all objects are created, the
init callback for each object that sets it is called.
Objects can lookup other objects in this callback.