Simplify core callbacks
The core passes a set of callback to each object in
order for the objects to be able to lookup objects,
complete commands, issue events, etc.

The way this was done is by passing a structure
containing all these callback to each object and have
the object manage it on its own. This was causing a lot
of replicated code. Also calling all of the callbacks
was clumsy.

Rework this by adding the structure containing all of
the callback to the base core_object_t type. After an
object is created the core will initialize this
structure.

To make calling all of the callbacks easier, a set of
macros are provided, as well.