Prevent step overshoot
The stepper object calculates how many steps to perform
on each update cycle. At the very end of the move, the
number of steps to perform could be larger than what is
left in the command. This can lead to step overshoot.

Fix this by limiting the number of steps done in the last
cycle.