Add API for allowing objects to submit commands
Sometimes, it is necessary for object to submit commands for
other objects. This also has to be done quickly and in a way
that is not blocking the submitting object (since submission
could be done in the update cycle.

To do this provide an API that allows objects to queue commands
and have the core execute commands in it worker thread.

This API does not currently allow scheduling of commands for
execution at a specific clock tick.