Provide a way for logging through Python
Intead of using printf statements inside the core code,
provide a way for all C code to log to the standard
Python logger.

This not only fixes all warnings about printf() but also
provides a unified way to log messages and adhere to
logging levels set on the command line.

Logging does affect the core timing, however, so it can
not be guaranteed if logging level is DEBUG or lower.