Add axis move and home commands
The axis "move" command issues a stepper move
command for the number of steps based on the
distance requested and ratio of distance to
steps.

The "home" command is a little bit more
complicated. It issues successive move commands
until the position is equal to the axis length
(the direction is based on the type of the
endstop associated with the axis).