Check for duplicate objects
Objects are created based on the their type and name.
Up to now, the core did not check whether there were
any duplicate objects.

This isn't necessarily a problem from the core's point
of view since it does not manage objects based on their
names. However, it can be confusing and hard to debug
because logging entries do not know the diffeence
between the entries.

It is also better to disallow duplicate entries since
that means the config file may have issues.