Remove rate-limiting logging
Rate-limiting logging was an effort to only display
log message at a certain rate. The use case was to
avoid swamping the console with update messages.

However, it never worked correctly since it required
a unique counter for each object that was created.
Since all objects of a certain class use the exact
same functions, the rate counter had to be part of
the object structure (so it can be counted per
object). This is not something that was OK to put
in the object structure.

As a result, remove the rate-limiting logging.