Fix timing test code
Originally, the timing test tool was a way to test
whether PThreads and C were fast enough to be able
to run contoller code with timing approaching the
clock rate of a STM32F446 chip.

It was also meant to test differences in priority
and scheduling of PThreads.

However, the error handling was not done correctly
leading to incorrect conclusions.

This change fixes the error handling. With this, it
seems that it is not possible to set priority and
scheduling policy for some reason.