Pass heap-allocated data when submitting events
The data structure passed to the core when submitting
event should be allocated on the heap since it will
outlive the function that submitted it.

If the data structure is allocated on the stack, it
will be free when the function returns, which could
lead to data corruption when the event is eventually
processed.