Implement object allocation cache
When objects submit command completions or events, the
associated data has to be allocated on the heap because
it has to be valid after the object function has
returned.

This means that objects have allocate memory using the
libc allocators. Since a lot of the command completions
or events can be submitted in the object update routines,
it could take extra time to do those allocations.

To improve performance, this change implements an object
cache mechanism similar to the one used in the Linux
kernel.

Objects can create caches of structures of different sizes.
The cache allocates a large number of such structures
upfront. After that, allocations only cost the time to pull
one of those structures off the list of pre-allocated
structures.