Switch to meson-python builds
Meson-python is a much more flexible build system
than setuptools. It allows for building the shared
libraries that we need, naming them correctly, and
installing them in the correct locations.

It also has a much easier way to build AND link to
a common shared library that provides utility
functions.