Handle non-cache objects in object_cache_free
Allocation of a cache object can happen at a different
place then freeing it. So, there would be no way for the
code freeing the memory to tell if it should be freed with
object_cache_free() or just free().

To handle this, enhance object_cache_free() to handle not
only NULL pointers (matching the behavior of free()) but
also heap-allocated memory. This is done by tagging cache
objects and determining how to free the memory based on the
tag and pointer values.