Fix stepper bug that would cause corruption
There was a timing race that would cause the stepper objects
to complete commands that had not been executed yet.

The stepper code records the currently executed command in a
pointer variable. When the command is complete, the values in
that command structure to complete the command.

However, if the update routine was called after the pointer was
set but before the command was executed/completed, the update
routine would complete the command. This was due to the update
routine not check what the actual command was.