Add support for absolute and relative coordinates
Adding support for both absolute and relative coordinates is done
by switching the axis "move" command to an absolute axis position.
Axis "move" commands now specify the absolute position along the
axis where the it should be at the end of the move.

With this change, relative coordinates can be implemented in the
frontends by querying the current axis position and computing the
new one based on the distance specified.