Change how probe trigger is computed
The probe update function used to set the 'triggered'
flag to false and then recompute if it should be
changed back to true.

Instead of doing that, change the value based on the
computation result.

This fixes an bouncing effect if the probe status is
read exactly between the setting of the flag to false
and setting it back to true if the probe is triggered.