Add a simple monitoring tool
It is hard monitoring the status of the emulation using
debug log entries. Those are useful for analyzing the
emulation from a logfile.

However, it is useful to monitor the state of all the
objects in the emulation.

For that purpose, this commit includes a simple Gtk tool
that continously monitors and reports the status of all
of the objects.