Unify frontend serial interface handling
Each of the frontends used to create and managed their
own serial interface. However, for the most part that
code was the same between all frontends.

To improve code duplication and maintainability, unify
the serial interface creation and event processing to
the base frontend class. Each frontend still processes
the received data separately.