Improvements to heater temperature interpolation
Normally, when heaters are set to heat to a temperature that
is close to the ambient, the heater reaches that temperature
very fast. The slowing of the temperature increase starts
happening as the target temperature approaches the max heater
temperature.

To better simulate that the heater object is enhanced to add
a maximum temperature setting, which is used to compute the
temperature ramp duration. The closer the target temperature is
to the maximum heater temperature, the longer the interpolation
lasts.

This is not perfect since it real-world applications, target
temperatures that a close to ambient (or the heater power is
high enough) will not experience any significant slow down as
the target temperature is approached. However, using the
interpolation duration is close enough.