Fix event handling from Python code
Event registration and handling in frontends was completely
broken since the event (un)register function was always set
to a dummy one. This was due to the fact that the function
were set in the frontend contructor but at that time, it
was not possible to pass anything to the contructor.

This commit fixes the event registration path. It also moves
the event decoding to the controller rather than the emulator.

As a result of this rework, it should now be possible for
virtual objects to send event notifications.