Don't hardcode object module names
The core had the names of the HW objects hard coded.
This makes adding new modules a little bit harder since
their name would have to be added to the array. This
step is, however, completely unnecessary since the name,
and thus the path, can be determined from the object name
array.