Fix bug with axis movement
Axis objects had a bug in the their move logic which would
cause the position to bounce around the target.

The reason for this bug was the distance computation. Previously,
if the remaining travel distance was less than a single motor step,
the distance was set to 1mm. This, of course, would cause the axis
to overshoot the target. It would then attempt to correct.

Instead, if the distance is less than one motor step, just adjust
the axis position by the remaining distance.