Per-object update threads
The update cycle used to done by a single update thread
that will go through all of the created objects and call
their update function.

This meant that object updates had to wait for each other
to be called.

Switching the object so they use their own thread allows
the to update independantly. This should improve object
update, and thus, emulation performance. However, it
seems that updates have to be extremely slim in order
to fit within controller ticks - a controller running
at 10MHz has an update cycle of 100ns.