Add framework for various kinematics support
The emulator should be able to support various kinematics models
to be able to represent different style machines.

This change adds a framework for supporting different kinematics.
The kinematics logic is implemented as a shared library so the
HW objects can use it to calculate positions. There is also a CFFI
wrapper that exposes the API to Python.