Decode byte array strings
The C helper function that parsed a ctype Structure
and returned it in the form of a dictionary set strings
to their bytearray form. That required that all users
of the function decode each used string in the
dictionary.

Instead of leaving the decoding to the users, just
decode all strings so they are ready for use.