Switch axis movement to be based on kinematics model
This is a fairly large change the puprose of which is to
switch how motor/axis movement is done in order to take
advantage of the kinematics models.

While previously movement was done using axis commands, it
is now done using motor commands. The axis objects update
their position based on the position of the motors. This
allows for the axis position to be updated based on the
kinematic model being used while the motor position is the
absolute value.

Achieving this required that the movement ratio parameter
be moved from the axis object to the motor/stepper object.

A new Python kinematics class is added to aid with movement
calculations and to provide a single interface to the
different kinematics models.

Note that the corexy kinematics model still does not work.
There is currently no way for the axis objects to accurately
update their position based on the position of two different
motors. For that to work, the axis objects would need a new
setting decoupling the name from the coordinate axis type.
This way, each axis object can use the coordinate axis type
to correctly set state for the axis position calculation.