Add a simple command response protocol
The emulator setups up a bidirectional pipe (/tmp/vortex)
that is used by clients to interface with the frontends.

While that allowed for clients to submit commands to the
frontends, there was no way for the client to get status
of the submitted commands.

This commit adds a simple response protocol that clients
can use to get replies from the frontends.

Not all frontends need to support this. Frontends that
interface with specific clients which implement their own
protocol don't need to use this.