Add a testing framework and tests
Development is becoming harder without the ability to
verify correctness. The lack of testing leads to more
bugs and regressions.

This commit adds a testing framework and a set of tests
that do basic verifications of all the existing modules
(with the exception of the 'probe' module, which needs
non-trivial rework first).

Using this framework, testing is automated and can be
run at arbitrary interval to verify correct operation.