Fix a small bug in stepper status reporting
The stepper object tracks steps as floating point numbers.
It does this so it can better emulate acceleration and
decelaration.

However, it can't report the step values as floating point
numbers as that doesn't make sense (for now, at lease since
there is no support for microstepping).

So, the values are not reported as rounded integers.