Convert toolhead to a HW object
This is step 1 of the rework of the probe object.

The probe object is tied to a toolhead not an axis.
Therefore, its state should be a function of the
toolhead position instead of the axis position. In
order for the probe object (which is a HW object)
to get the toolhead's state, the toolhead also needs
to be a HW object.

This is because there is no way for a HW object to
query the state of a virtual object. Even if such a
path could be added, it would require taking the
Python GIL, which is a performance killer.