Rework probe object
The probe object should be tied to a toolhead not an
axis since most of the times, probes are attached to
a toolhead at a certain set of offsets.

This commit implements this by reworking how the
probe object works. It now uses the status from the
toolhead it is attached to in order to determine
its own status.

This change also allows for the probe test to be added.