Process work in the core in batches
Due to a bug in the endstop object, the object would
continuously generate 'triggered' events. This would
end up creating a very long list of events which needed
to be processes.

This lead to all other work to be left waiting behind
all of the events.

This commit implements a batched processing mechanism
where each type of work - events, command submission, and
completions - are processed in batches. This way there is
no possibility for any of the types of work to block the
entire thread.