Yet another rework of controller timing
It is not reasonable to expect that each object's
update loop will execute in the amount of time allowed
based on the controller frequency. This is more true
with higher frequencies. This inability to keep precise
timing will throw off any clients that may depend on it.

Instead use the absolute time to calculate the number of
controller clock ticks based on the specified frequency.
Controller runtime is still measured in nanoseconds for
any objects that depend on absolute wall clock time.

This change also correctly plumbs the controller tick and
runtime methods through the Python extension calls. It also
supplies the controller methods to the frontends.