Correct implementation of emulation frequency
The emulation frequency command line argument was to
be used as means to slow down or change the controller
frequency. However, that does not actually provide an
accurate emulation.

Controllers run at specific frequencies based on their
hardware. It has nothing to do with the emulation
frequency - the latency of the object update cycle.

So, now the command line argument is used to specify
the update latency, while allowing the controllers
to run at their specified frequency.