Don't use the emulator's logfile option for tests
Up to now, the tests framework used the emulator's own
log file support if a log was required. However, that
does not allow the framework to insert its own log entries
in the log.

It is more useful for the framework to be able to insert
messages into the log stream coming out of the emulator.
This makes debugging much easier. So, instead of using
the emulator's log file, capture the debug output and
write it to a logfile from the framework.