Atomically exchange the command ID on completion
There are some HW objects which record and use the current
command/command ID in order to check if a command is being
executed. After the command is complete, the HW object sets
the recorded command ID to 0.

This leaves a small window during which another thread can
try to submit a command, which will be rejected due to the
recorded ID not having been cleared.

To fix this, use an atomic exchange operation during command
completion which will reset the command ID.