Synchronize state update and state report
The probe object needs to perform a number of check in
order to update its status. A state query can come in
at any time during those tests and get a inconsistent
state.

To avoid such a situation, wrap the critical sections
of both code paths in a lock which will guarantee a
consistent view of the state.