Improve core thread control before adding timers
This commit is a bit of a mess - the goal was to improve
the core thread control code in preparation for adding timer
support. In the process, the Python extension module got
re-organized and the Python class methods renamed.

So, the commit is much larger that it should have been but
it turned out that untangling all of the changes was a lot of
work.