Add controller core timers
This change adds timer support to the controller core.
The timer APIs allow other emulator entities to register
and schedule periodic timers, which are a function of
the controller's clock frequency.

The first usage of the timers should be the Klipper
frontend, which schedules events and queries on a periodic
basis.