Move command queue to frontends
Move the command queue to the frontend rather than the top emulator
module. This would allow for each frontend to use their own command
queue that suites the frontend better.

This commit also removes the ScheduledQueue since that is not needed.
The intent was for it to be used by the Klipper frontend since Klipper
schedules commands at specific clock times. However, it makes much more
sense for Klipper to submit the commands at the appropriate time rather
than use a special queue type.