Expose stepper pins through a shared memory location
The work with the Klipper frontend exposed a flaw with how object
commands are implemented.

Klipper requires tight timing when performing motor moves and it
controller when the steps are made using timers that trigger at
specific clock ticks.

Using the established method for submitting object commands had two
major flaws:
   1. It is hard to reliably execute a single command at a time,
      which causes command execution errors and missed steps.
   2. The command execution path was way too slow to support the
      required timing.

This meant that steppers could not be moved successfully with the
Klipper frontend.

To fix this, expose a memory location that serves as stepper pin
control. The stepper object monitors the value in the memory
location and performs actions based on psuedo-pin values. This
enables the stepper object to update it's state much faster and,
in turn, allow Klipper to successfully move the motor.