Add a tool to generate an emulator config from Klipper config
One of the main pain points when using the Klipper frontend is routing
pins to the appropriate emulator objects. To make this easier, this
tool will generate an emulator config file from the Klipper config file
setting the emulator object pins to correspond to the pins defined in
the Klipper configuration.