Fix C code indentation
I've been having a lot of trouble getting the C code formatted
properly. The Emacs indentation has not worked correctly resulting
in incorrect indentation of a lot of the code.

This is a large change that attempts to fix this by fixing the
Emacs settings in .dirs-local.el, adding a .clang-format file, and
re-indenting the entire C code base.