Provide a more direct method for querying the endstop status
Provide a way to query endstop trigger status through a pin word
like the stepper control word. This should make it faster to
report triggered status.

Along with the above, the word addresses for the endstops, probes,
and steppers are added to the status structures. This way the
endstop and probe objects don't need to add a separate command for
getting the word address.