Switch heater to a genuine temperature simulation
The heater object has been using a very crude temperature
algorithm which just created an gradual curve. This algorithm
was not based on any actual thermal math, just a simple
easing function.

This commit changes this by switching the algorithm to an
actual thermal simulation. The algorithm is based on
Paniel Detersson's printer heated bed simulator. It has
been converted and incorporated into the Vortex emulator
with his permission.