Order reads and write to shared variables
The thread control logic uses shared variables between
threads for both control and data. However, there was
nothing there to make sure that read/write ordering is
maintained.

In order to provide some ordering semantics, we can use
the 'volatile' keyword, which guarantees that accesses
to shared variables are order as they appear in the
source code.