Fix axis position initialization
The axis starting position is randomized in order to
emulate real-world situations where the machine is
lest/powered up in a unknown state.

The random axis position was generated by just getting
a random number that is within the limits of the axis
length.

However, any axis position has to be a function of the
stepper motor positions and, therefore, the axis
position has to be a multiple of the stepper motor
steps.

In order to achieve this, the random number is used as
a starting point and then adjusted to be a multiple of
the stepper's steps.