Add support for nested object configuration structures
This commit adds support for parsing and filling nexted ctypes
structures in object configurations. With this an object
configuration structure can have substructures are members. The
configuration file syntax adds a new way of parsing settings, as
well.

There is no reason why this new support cannot be used for all
object structures, not only configuration.