Report and use update and controller frequencies differently
Controller objects reported the controller frequency to frontends
but not update frequency.

In the case of the Klipper frontend knowing the update frequency
is useful since it can cause errors if the update frequency is too
low. Klipper relies a lot on timers to processes commands an
responses. If the update frequency is too low, the timer callbacks
get called with a higher clock difference, which causes the
Klipper errors.

Knowning the update frequency allows the Klipper frontend to
report a warning.