Fix stepper step tracking on high speed moves
High speed moves done by the Klipper frontend did not get
processed correctly resulting in the stepper moving to the
wrong coordinate. This was due to the stepper thread not
being able to catch all step pulses from the frontend.

To fix this, the low 16 bits of the pin word are turned
into a step counter that is incremented by the frontend
and reset to 0 by the HW object. In addition, the pin
word is manipulated only using atomic operations to
ensure that read-modify-write operations by the frontend
and HW object don't interleave each other.