Add tool to help with printing Gcode files
This simple tool reads the Klipper config file looking for
the '[virtual_sdcard]' section. From there, it extracts the
path where the Gcode files are stored and displays them for
selection.

Once a file is selected, it submits the Klipper commands
required to have Klipper start "printing" the Gcode file.