Add debug tool to visualize toolhead travel
This tool monitors are records the toolhead's position at very
short intervals. Upon interruption by the user, the tool creates
a 3D scatter plot of all the toolhead positions.

The intent is to help visualize (and thus, verify) that the
emulator is correctly moving the toolhead based on the input
from the frontend.