Implement Klipper 'FIRMWARE_RESET' command
Handling of the 'FIRMWARE_RESET' Klipper command requires
the handling of the 'reset' protocol command handler.

In the case of the Klipper frontend implementation, this
required overriding of the 'reset' BaseFrontend method.
However, the correct class method was not being found when
the Klipper frontend was looking it up because the base
class defined it as an attribute. To fix this, the base
class has to define the controller methods are attributes
but also provide methods that call those attributes.

Also, helper 'shutdown' methods needed updating to handle
calls to their 'shutdown' method followed by calls to the
'__del__' method.