Fix Klipper uptime reporting
The Klipper frontend was not accurately reporting the 'uptime'
of the emulator. The value that were being returned were
incorrect, resulting in Klipper host to get continuously
confused about the frequency of the controller.

With this fix, Klipper does not end up continuously adjusting
the clock prediction values, which results in a more accurate
match in timing between the emulator and Klipper host.