Fix Stepper helper step timing
The Stepper helper stil had issues with maitaining
correct step timing. This resulted in a large mismatch
between emulator and host with respect to the step
position of a stepper motor.

This commit improve the timing of the helper. This
results in a much better match between the emulator
and Klipper