Use a single lock for managing timers
Timers used one lock per timer list when managing the
timers. The intent was to provide as much concurrency
as possible to make timer code faster.

However, the multiple locks were causing issues when
the timers were being moved from one list to the other.
To fix this, the timer code now uses a single lock.

This does make it more likely to have lock contention
and potentially have threads waiting but it simplifies
the logic and improves correctness.

This commit also adds some timer debugging code that
was used to debug the various issues.