Improve timing of Klipper helpers
This commit has a number of improvements to several Klipper
helper classes meant to improve object timing and handling.

The biggest change has gone into the Stepper helper. It not
only improve step timing but also includes some "synchronization"
code that ensures that steps are done on time and executed
before direction is changed.