Add stepper debug pin
Add a debug control pin to stepper objects. This pin can be used
to signal the stepper controller object to perform any debug
actions.

Also, remove the stepper pin thread sleep. The stepper pin thread
should run as fast as possible to ensure that all steps are
processed as quickly as possible.