Add tool to generate Klipper protocol data
The Klipper frontend uses a set of namespaces that describe
the Klipper host-MCU protocol.

Previously, the file containting the protocol description was
generated using a tool that produced somewhat correct code.
Then, that code was edited manually to produce the correct
protocol data.

The tool used regular expressions to try and find the Klipper
command/response pairs but it failed to match the response to
the command in all cases.

The new tool being added with this commit uses Clang to generate
a C AST from the Klipper source files. This allows the tool to
get extra metadata that helps it match commands to responses.

Using the C AST, the tool is able to generated a complete and,
more importantly, correct protocol data without the need for
manual editing. The main advantage of this is that the protocol
can be easily regenerated if it changes in the future.