Prevent a double stepper direction flip
There was a bug in the Klipper stepper helper that was causing
a double direction flip in some cases:

  1. `queue_steps` command would be received and the helper would
     add it to the queue.
     1.1 If the current move's move count is 0, the next step
         would be loaded. This is done in order to kick off the
         step generation.
  2. After the set of steps has been queued, the timer-based
     callback takes over the execution of the tests. Any additional
     moves are just queued on the move queue.

There was a race condition where both `Stepper::queue_move` and
`Stepper::_next_move` would change the stepper direction:

   1. The stepper is completing the last step of a move. The
      timer callback has been called and has executed the last
      step. At this time, the current move's step count is 0.
   2. The callback loads the next move, which requires a
      direction change.
   2. `Stepper::_next_move` changes the stepper direction.
   3. Prior to `Stepper::_next_move` resetting the current move
      object, `Stepper::queue_move` is called to add a new move
      to the queue. This new move requires a direction change,
      as well. Since the current's move step count is 0, the function
      calls `Stepper::_next_move`, which ends up changing the
      stepper direction.

At this point, the stepper's direction has not only been inverted
but an entire move has been skipped because `Stepper::queue_move`
would end up resetting the current move with the values of the
newly queued move.

To fix this, the current move object get a new property - `active`.
`Stepper::queue_move` only calls `Stepper::_next_move` is the
current move is not in the active state. The active state is
disabled only when current move is complete are there are no more
moves in the queue.