Use nanosleep in time control loop
The time control loop was previously using `clock_nanosleep()` to
sleep between iterations. However, the call to `clock_nanosleep()`
was trying to use `CLOCK_MONOTONIC_RAW` to sleep against.

`CLOCK_MONOTONIC_RAW` is not a valid clock that `clock_nanosleep()`
accepts. Therefore, the call returned immediately with an error.
Since the call was not error checked, this resulted in the time
control loop to run without the intended sleep, which resulted in
it consuming 100% of CPU cycles.

The reason to use `CLOCK_MONOTONIC_RAW` was to guard against
discontinuous clock jumps. However, according to documentation of
`nanosleep()` it is protected against that.