Small speed up of heater update function
The heater update function needs to record that start of the
computation cycle in order to correctly compute the change in
temperature between updates. This needs to be done at the
beginning and needs to be done only once.

Instead of using a branch, using atomic compare/exchange, which
accomplishes the same thing but using a single atomic instruction.