Use `SDCARD_PRINT_FILE` in favor of `M23`
The Klipper GCode parser seems to be having issues with parsing
filenames that start with numbers. For example, issuing the gcode
`M23 3DBency.gcode` results in it attempting to execute `M23 3`.

Instead using `SDCARD_PRINT_FILE` which has a command argument
used to specify the GCode file and, thus, avoids the issue.