Determine CPU frequency in a better way
Rather than try to parse the CPU module name for its
frequency (which isn't reliable), use the 'cpu Mhz'
value that is in /proc/cpuinfo.

Parsing the CPU module name is not reliable because
not all model names end with the CPU's speed. As an
example, the module name reported on an Alienware
m16 R2 is "Intel(R) Core(TM) Ultra 7 155H".

While debugging this a bug in `parse_frequency()`
was found with the regular expression used to parse
the frequency string.