Allow the user to decide if the emulator should set thread priority
Expose the ability to set emulator thread priority to the user through
a command line option. This not only gives more flixibility but makes
error reporting easier as the error can be propagated up.

Another improvement is to atomically set the thread running state so
if the threads are stopped prior to their loops starting, they can
properly exit.