Correctly handle timer removal during execution
A condition existed where if a timer was removed during the time
its callback is being run, it could result in a use-after-free
situation as the timer structure would have been freen when the
callback returns.

Handle this correctly by checking if the timer is currently
executing in the unregister function and if so, just signal the
timer execution function that the timer has been removed.