Enable finer-grain log filtering
This has been something that was needed for a long time in
order to help with debug/development.

This commit greatly restructures the logging support for
the emulator. The code now uses a hierarchy of loggers.
This enables the ability to filter messages from specific
parts of the emulator.

With this support, filters can be setup so messages only
from specific object and/or modules can be displayed,
greatly reducing the noise.