Rework how emulation elements are setup
The way frontends get information about the controller
element was pretty clunky and hard to extend.

This commit changes all that by changing how the elements
are passed around by the top Emulator object. This allows
frontends to have direct access to the controller which
makes querying parameters much easier.