Fix Klipper shutdown handling
Handling of Klipper commands while the emulation is in
the "shutdown" state was corrent if Klipper sent any other
command than "restart".

With the "restart" command, there is no protocol object for
the command. It's just a string, which maps to a method that
is part of the frontend object. This caused an exception
because the frontend has trying to treat the command string
as a protocol object.