Add dprint() debugging function
This new function combines the printing of the message and
flushing of the stdout stream, which should improve debug
output accuracy when there is output from multiple threads.

This new function is intended for one-off debug messages
that will not be part of the actual code base. For permenant
message, the existing logging facilities should be used.

Note that the entire emulation will be slowed down since
every single debug message will cause flushing of the stream.