Set RPATH correctly and fix build errors/warnings
Change the Meson build options to correctly set the RPATH
in shared object files. For now, linker flags are used,
even though Meson warn that this will become an eror in
the future. However, Meson does not support setting multiple
paths with the `intall_rpath` option.

Also, fix some compile errors/warning that we found during
testing of the above.