Handle multiple objects of a klass through a proxy
In some cases (like display objects), it might be useful to
implement support for different virtual object subtypes as
subclasses of a common virtual object class. What this would
mean is that there would be multiple classes with the same
object type.

This would not have worked since there is no way to figure
out which subclass to use. What would end up happening is that
the latest subclass that was inspected would be the one used
for the emulation.

With this commit, using multiple subclasses is made possible
by requiring a "factory" function, which can use some
configuration to determine which of the subclasses should be
used.