Separate motor count check for CoreXY and CoreXZ kinematics
During object creation, the axis object performs some checks and
sets properties based on the kinematics model used for the
emulation.

In the case of CoreXY, CoreXZ, and Cartesian kinematics, a lot of
these are common between the 3 types. So, to reduce code duplication,
the implementaion uses a `switch` statement where the CoreXY and
CoreXZ cases fall through to the Cartesian one.

Due to the way the checks where written, there is a bug when the
kinematics model is CoreXY:

  1. The CoreXY case would check that the number of motors is 2
     if the axis type is either `AXIS_TYPE_X` or `AXIS_TYPE_Y`.
     If successful, it would fall through.
  2. The CoreXZ case would check that the number of motors is 2
     if the axis type is either `AXIS_TYPE_X` or `AXIS_TYPE_Z`.
     If not, it will cause an error.
  3. If the model is CoreXY and the axis type is `AXIS_TYPE_Y`,
     the first check will succeed but the second would fail.

To fix this, explicitly check if the model is CoreXZ in the
CoreXZ case. This will cause the check to be skipped if the model
is CoreXY.