Split command, event, and completion processing
Instead of having one large function which does all of the
processing for submitting command, trigerring event callbacks,
and returning command completions, split it into separate
functions and create separate threads for them.

Having a single function meant that there had to be a batch
size for each of the three parts so one does not take too much
time. This is now unnecessary with a dedicated function/thread.